Revert "Test by changing signatures from bytearray to int arrays"
This reverts commit a3c4635a66c91be52cd0ee8b54733f30be89d6bb.

+++ LDLSH/src/main/java/SystemLayer/Data/LSHHashImpl/LSHHash.java
@@ -17,7 +17,7 @@ public interface LSHHash extends Serializable, Comparable<LSHHash> {
      * Returns the generated complete signature
      * @return a bite array containing the associated object's signature
      */
-    int[] getSignature();
+    byte[] getSignature();
 
     /**
      * Returns the list of signature blocks
